refactoring generar_filas_columnas
